# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Claude Code container orchestration API - secure, isolated AI execution",
        "title": "Stromboli API",
        "termsOfService": "https://github.com/tomblanc/stromboli",
        "contact": {
            "name": "API Support",
            "url": "https://stromboli/issues"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/images": {
            "get": {
                "description": "Returns all local container images sorted by compatibility rank. Images with rank 1-2 are verified compatible, rank 3 is standard glibc (compatible), rank 4 is incompatible (Alpine/musl).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "images"
                ],
                "summary": "List images",
                "responses": {
                    "200": {
                        "description": "List of images (empty array if none exist)",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ImagesListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/images/pull": {
            "post": {
                "description": "Pulls a container image from a registry. This operation may take some time for large images.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Pull image",
                "parameters": [
                    {
                        "description": "Pull request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.ImagePullRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Image pulled successfully",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ImagePullResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Pull failed",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/images/search": {
            "get": {
                "description": "Searches container registries for images matching the query. Returns results from Docker Hub and other configured registries.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Search images",
                "parameters": [
                    {
                        "type": "string",
                        "example": "python",
                        "description": "Search query",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "example": 10,
                        "description": "Maximum number of results (default 25, max 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "example": true,
                        "description": "Don't truncate output (show full descriptions)",
                        "name": "no_trunc",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ImageSearchResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request (missing query)",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Registry search failed",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/images/{name}": {
            "get": {
                "description": "Returns detailed information about a specific container image including all labels and compatibility information.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Inspect image",
                "parameters": [
                    {
                        "type": "string",
                        "example": "python:3.12-slim",
                        "description": "Image name with optional tag",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Image details",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ImageDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid image name",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Image not found",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/jobs/{id}": {
            "get": {
                "description": "Returns the status and result of an async job",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Get job status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api.JobResponse"
                        }
                    },
                    "404": {
                        "description": "Job not found",
                        "schema": {
                            "$ref": "#/definitions/internal_api.RunResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Cancels a pending or running job",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Cancel job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Job not found",
                        "schema": {
                            "$ref": "#/definitions/internal_api.RunResponse"
                        }
                    },
                    "409": {
                        "description": "Job cannot be cancelled",
                        "schema": {
                            "$ref": "#/definitions/internal_api.RunResponse"
                        }
                    }
                }
            }
        },
        "/run": {
            "post": {
                "description": "Executes Claude Code in an isolated Podman container",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "execution"
                ],
                "summary": "Run Claude",
                "parameters": [
                    {
                        "description": "Run request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.RunRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api.RunResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/internal_api.RunResponse"
                        }
                    },
                    "500": {
                        "description": "Execution failed",
                        "schema": {
                            "$ref": "#/definitions/internal_api.RunResponse"
                        }
                    },
                    "503": {
                        "description": "Claude not configured",
                        "schema": {
                            "$ref": "#/definitions/internal_api.RunResponse"
                        }
                    }
                }
            }
        },
        "/run/async": {
            "post": {
                "description": "Starts Claude Code execution asynchronously and returns a job ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "execution"
                ],
                "summary": "Run Claude async",
                "parameters": [
                    {
                        "description": "Run request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.RunRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/internal_api.AsyncRunResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/internal_api.RunResponse"
                        }
                    },
                    "503": {
                        "description": "Claude not configured",
                        "schema": {
                            "$ref": "#/definitions/internal_api.RunResponse"
                        }
                    }
                }
            }
        },
        "/run/stream": {
            "get": {
                "description": "Executes Claude and streams output in real-time using SSE",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "execution"
                ],
                "summary": "Stream Claude output",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The prompt to send to Claude",
                        "name": "prompt",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Working directory inside container",
                        "name": "workdir",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Session ID for conversation continuation",
                        "name": "session_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event stream of output lines",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "503": {
                        "description": "Claude not configured",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/secrets": {
            "get": {
                "description": "Returns metadata for all available Podman secrets that can be injected into agents. Secret values are never returned - only IDs, names, and creation times.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "secrets"
                ],
                "summary": "List secrets",
                "responses": {
                    "200": {
                        "description": "List of secrets (empty array if none exist)",
                        "schema": {
                            "$ref": "#/definitions/internal_api.SecretsListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.SecretsListResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new Podman secret that can be injected into agents. Secret names must be alphanumeric (with dashes and underscores), max 253 characters. Values are limited to 1MB.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "secrets"
                ],
                "summary": "Create secret",
                "parameters": [
                    {
                        "description": "Create secret request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.CreateSecretRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Secret created successfully",
                        "schema": {
                            "$ref": "#/definitions/internal_api.CreateSecretResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request (missing/invalid name or value)",
                        "schema": {
                            "$ref": "#/definitions/internal_api.CreateSecretResponse"
                        }
                    },
                    "409": {
                        "description": "Secret with this name already exists",
                        "schema": {
                            "$ref": "#/definitions/internal_api.CreateSecretResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.CreateSecretResponse"
                        }
                    }
                }
            }
        },
        "/secrets/{name}": {
            "get": {
                "description": "Returns metadata about a specific Podman secret. For security, the actual secret value is never returned - only the ID, name, and creation time.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "secrets"
                ],
                "summary": "Get secret metadata",
                "parameters": [
                    {
                        "type": "string",
                        "example": "github-token",
                        "description": "Secret name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Secret metadata",
                        "schema": {
                            "$ref": "#/definitions/internal_api.SecretInfoResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid secret name",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Secret not found",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Permanently deletes a Podman secret. This action cannot be undone. Secrets currently in use by running containers may cause those containers to fail.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "secrets"
                ],
                "summary": "Delete secret",
                "parameters": [
                    {
                        "type": "string",
                        "example": "github-token",
                        "description": "Secret name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Secret deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/internal_api.DeleteSecretResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid secret name",
                        "schema": {
                            "$ref": "#/definitions/internal_api.DeleteSecretResponse"
                        }
                    },
                    "404": {
                        "description": "Secret not found",
                        "schema": {
                            "$ref": "#/definitions/internal_api.DeleteSecretResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.DeleteSecretResponse"
                        }
                    }
                }
            }
        },
        "/sessions": {
            "get": {
                "description": "Returns all existing session IDs",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "List sessions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api.SessionListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.SessionListResponse"
                        }
                    }
                }
            }
        },
        "/sessions/{id}": {
            "delete": {
                "description": "Removes a session and all its stored data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sessions"
                ],
                "summary": "Destroy session",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api.SessionDestroyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api.SessionDestroyResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api.SessionDestroyResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "internal_api.ImagePullRequest": {
            "description": "Request to pull a container image from a registry",
            "type": "object",
            "required": [
                "image"
            ],
            "properties": {
                "image": {
                    "type": "string",
                    "example": "python:3.12-slim"
                },
                "platform": {
                    "type": "string",
                    "example": "linux/amd64"
                },
                "quiet": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "internal_api.ImagePullResponse": {
            "description": "Result of image pull operation",
            "type": "object",
            "properties": {
                "image": {
                    "type": "string",
                    "example": "python:3.12-slim"
                },
                "image_id": {
                    "type": "string",
                    "example": "sha256:abc123def456"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "internal_api.ImagesListResponse": {
            "description": "List of local container images sorted by compatibility",
            "type": "object",
            "properties": {
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_api.ImageInfoResponse"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "OAuth2 Bearer token (future)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}